changing count from u16 to usize
I don't think there was any real reason for that to be u16 in the first
place

+++ src/lib.rs
@@ -124,7 +124,7 @@ pub struct User {
     pub name: String,
 
     #[serde(default = "get_zero")]
-    pub count: u16,
+    pub count: usize,
 }
 
 fn empty_details() -> Details {
@@ -135,7 +135,7 @@ const fn empty_string() -> String {
     String::new()
 }
 
-const fn get_zero() -> u16 {
+const fn get_zero() -> usize {
     0
 }
